Extracted all visual elements to UI folder
