+++ app/ui/components/LogoHeader/index.js
@@ -5,8 +5,8 @@
  */
 
 import React from 'react';
-import Logo from 'images/logo.png';
 import styled from 'styled-components';
+import Logo from '../../images/logo.png';
 
 const StyledLogoHeader = styled.div`
   padding-top: 40px;
