+++ app/ui/components/VerticalContainer/index.js
